many cpu optimisations

+++ pkg/gb/ppu.go
@@ -53,7 +53,7 @@ const (
 
 // Set the status of the LCD based on the current state of memory.
 func (gb *Gameboy) setLCDStatus() {
-	status := gb.Memory.ReadHighRam(0xFF41)
+	status := byte(gb.Memory.ReadHighRam(0xFF41))
 
 	if !gb.isLCDEnabled() {
 		// set the screen to white
@@ -128,12 +128,12 @@ func (gb *Gameboy) setLCDStatus() {
 
 // Checks if the LCD is enabled by examining 0xFF40.
 func (gb *Gameboy) isLCDEnabled() bool {
-	return bits.Test(gb.Memory.ReadHighRam(LCDC), 7)
+	return bits.Test(byte(gb.Memory.ReadHighRam(LCDC)), 7)
 }
 
 // Draw a single scanline to the graphics output.
 func (gb *Gameboy) drawScanline(scanline int) {
-	control := gb.Memory.ReadHighRam(LCDC)
+	control := byte(gb.Memory.ReadHighRam(LCDC))
 
 	// LCDC bit 0 clears tiles on DMG but controls priority on CGB.
 	if (gb.IsCGB() || bits.Test(control, 0)) && !gb.Debug.HideBackground {
@@ -156,7 +156,7 @@ func (gb *Gameboy) getTileSettings(lcdControl byte, windowY byte) (
 
 	if bits.Test(lcdControl, 5) {
 		// Is current scanline we're drawing within windows Y position?
-		if windowY <= gb.Memory.ReadHighRam(0xFF44) {
+		if windowY <= byte(gb.Memory.ReadHighRam(0xFF44)) {
 			usingWindow = true
 		}
 	}
@@ -182,10 +182,10 @@ func (gb *Gameboy) getTileSettings(lcdControl byte, windowY byte) (
 // Render a scanline of the tile map to the graphics output based
 // on the state of the lcdControl register.
 func (gb *Gameboy) renderTiles(lcdControl byte, scanline int) {
-	scrollY := gb.Memory.ReadHighRam(0xFF42)
-	scrollX := gb.Memory.ReadHighRam(0xFF43)
-	windowY := gb.Memory.ReadHighRam(0xFF4A)
-	windowX := gb.Memory.ReadHighRam(0xFF4B) - 7
+	scrollY := byte(gb.Memory.ReadHighRam(0xFF42))
+	scrollX := byte(gb.Memory.ReadHighRam(0xFF43))
+	windowY := byte(gb.Memory.ReadHighRam(0xFF4A))
+	windowX := byte(gb.Memory.ReadHighRam(0xFF4B) - 7)
 
 	usingWindow, unsigned, tileData, backgroundMemory := gb.getTileSettings(lcdControl, windowY)
 
@@ -201,7 +201,7 @@ func (gb *Gameboy) renderTiles(lcdControl byte, scanline int) {
 	var tileRow = uint16(yPos/8) * 32
 
 	// Load the palette which will be used to draw the tiles
-	var palette = gb.Memory.ReadHighRam(0xFF47)
+	var palette = byte(gb.Memory.ReadHighRam(0xFF47))
 
 	// start drawing the 160 horizontal pixels for this scanline
 	gb.tileScanline = [160]uint8{}
@@ -299,17 +299,17 @@ func (gb *Gameboy) renderSprites(lcdControl byte, scanline int32) {
 	}
 
 	// Load the two palettes which sprites can be drawn in
-	var palette1 = gb.Memory.ReadHighRam(0xFF48)
-	var palette2 = gb.Memory.ReadHighRam(0xFF49)
+	var palette1 = byte(gb.Memory.ReadHighRam(0xFF48))
+	var palette2 = byte(gb.Memory.ReadHighRam(0xFF49))
 
 	var minx [ScreenWidth]int32
 	var lineSprites = 0
 	for sprite := uint16(0); sprite < 40; sprite++ {
 		// Load sprite data from memory.
-		index := sprite * 4
+		index := uint(sprite * 4)
 
 		// If this is true the scanline is out of the area we care about
-		yPos := int32(gb.Memory.Read(uint16(0xFE00+index))) - 16
+		yPos := int32(gb.Memory.Read(0xFE00+index)) - 16
 		if scanline < yPos || scanline >= (yPos+ySize) {
 			continue
 		}
@@ -320,9 +320,9 @@ func (gb *Gameboy) renderSprites(lcdControl byte, scanline int32) {
 		}
 		lineSprites++
 
-		xPos := int32(gb.Memory.Read(uint16(0xFE00+index+1))) - 8
-		tileLocation := gb.Memory.Read(uint16(0xFE00 + index + 2))
-		attributes := gb.Memory.Read(uint16(0xFE00 + index + 3))
+		xPos := int32(gb.Memory.Read(0xFE00+index+1)) - 8
+		tileLocation := gb.Memory.Read(0xFE00 + index + 2)
+		attributes := byte(gb.Memory.Read(0xFE00 + index + 3))
 
 		yFlip := bits.Test(attributes, 6)
 		xFlip := bits.Test(attributes, 5)
